Made extrac_title_and_body part of rules file
